+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Version 4.2**
 
  - Fixed a bug in the dictionary validation that would not flag as invalid a dictionary with a missing URI attribute.
+ - Added support in NaaccrXmlUtils for new lineToPatient and patientToLine methods.
 
 **Version 4.1**